+++ src/pretalx/orga/views/cfp.py
@@ -307,7 +307,9 @@ def question_move(request, pk, up=True):
     if not request.user.has_perm("orga.edit_question", question):
         messages.error(request, _("Sorry, you are not allowed to reorder questions."))
         return
-    questions = list(request.event.questions.order_by("position"))
+    questions = list(
+        request.event.questions(manager="all_objects").order_by("position")
+    )
 
     index = questions.index(question)
     if index != 0 and up: